Increase precisioun by using double instead of float

+++ SupplyAndDemand/Entity.cpp
@@ -1,6 +1,6 @@
 #include "Entity.h"
 
-void Entity::Update(const float deltaTime, const float deltaHours)
+void Entity::Update(const double deltaTime, const double deltaHours)
 {
 
 }
